feat: value type for IP netmasks
This changeset is the first in a series to replace this module's IP
address types with those from netip. A value type to represented IP netmasks
following the pattern of netip.Addr, intending to replace the existing
IPMask type.

Unlike net.IPMask, the type introduced by the netmask package is
comparable (you can use `==` and use it as a map key), as well as being
immutable while taking up less memory, especially for IPv6.

Unlike netip.Prefix, this type does not require canonical prefix-based
netmasks for IPv4, nor are Mask directly linked to addresses.