test(client): run e2e test against a UML kernel
This package has tests of encoding and decoding netlink messages,
derived from real observed messages, there were no tests to confirm the
package worked against a running kernel.

This changeset adds an initial end-to-end test, confirming this package
can program a service and destination. It also adds a program compatible
with `go run` and `go test` to run the tets in a User Mode Linux kernel.