fix(netmask): text marshaler with multibyte prefix
For the IPv6 family the Mask type stores a prefix, rather than a full
128-bit mask. To implement the TextMarshaller interface for these
prefixes, Mask marshals the number of one-bits in the prefix to a
string.

However, the unmarshal function expected IPv6 family prefixes to only
ever be 1 byte long. This holds true for the BinaryMarshaller
implementation, but in text prefixes larger than 9 would be represented
by multiple bytes, one for each numeric digit.

This implementation adjusts the unmarshal implementation to accept
lengths up to 3 bytes for IPv6.

Fixes: bc6bb04 ("feat: value type for IP netmasks")