feat(netmask): implement new encoding interfaces
Go 1.24 introduces two new interfaces to the encoding package, which
have been implemented in net/netip: BinaryAppender and TextAppender.
These interfaces are similar to the BinaryMarshaler and TextMarshaler
interfaces, respectfully, but instead of creating a new byte-slice they
append to a provided slice.

As the two interfaces are closely related, this changeset implements
the *Marshaler interfaces in terms of the *Appender interfaces.